+++ src/com/landawn/abacus/util/ImmutableSet.java
@@ -16,8 +16,10 @@ package com.landawn.abacus.util;
 
 import java.util.Collection;
 import java.util.Collections;
+import java.util.HashSet;
 import java.util.LinkedHashSet;
 import java.util.Set;
+import java.util.SortedSet;
 
 /**
  * 
@@ -72,6 +74,6 @@ public class ImmutableSet<E> extends ImmutableCollection<E> implements Set<E> {
             return empty();
         }
 
-        return new ImmutableSet<>(new LinkedHashSet<>(set));
+        return new ImmutableSet<>((set instanceof LinkedHashSet || set instanceof SortedSet) ? new LinkedHashSet<>(set) : new HashSet<>(set));
     }
 }